Split BedSearchService by creating new CAS3 Bedspace Search Service
